API endpoints are implemented, exception handling added, TODO:unit tests,security

+++ src/main/java/com/bookmanagement/api/repository/BookRepository.java
@@ -4,6 +4,6 @@ import com.bookmanagement.api.model.BookEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BookRepository extends JpaRepository<BookEntity, Long> {
+public interface BookRepository extends JpaRepository<BookEntity, Integer> {
 
 }
